+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [2.1.0] - 2020-02-02
 ### Added
 
 - Termshark now provides a conversations view for the most common conversation types.
@@ -72,7 +72,8 @@ termshark -i eth0 --tty=/dev/pts/5
 
 - Initial release.
 
-[Unreleased]: https://github.com/gcla/termshark/commpare/v2.0.3...HEAD
+[Unreleased]: https://github.com/gcla/termshark/commpare/v2.1.0...HEAD
 [1.0.0]: https://github.com/gcla/termshark/releases/tag/v1.0.0
 [2.0.0]: https://github.com/gcla/termshark/releases/tag/v2.0.0
 [2.0.3]: https://github.com/gcla/termshark/releases/tag/v2.0.3
+[2.1.0]: https://github.com/gcla/termshark/releases/tag/v2.1.0